[FIX] Extract bootstrap class
Extract/move temporary class into file of its own to address coding-
standards.

Additionally provide backtrace with --verbose and similar command-line-
options when an exception is thrown when bootstrapping the application.
